Turned off serving tests in a GitHub CI
Committed on branch 'nginx'

+++ Frontend/FriendsWeb/tests/server_fixtures.py
@@ -8,6 +8,7 @@ from werkzeug.serving import make_server
 from base_dir import MANAGER_WORKDIR
 from config import WERKZEUG_TEST_PORT
 from helpers.osdir import OSDIR
+from helpers.context import CONTEXT
 from friends.main import APPLICATION as app
 
 from servers.nginxd import configuration
@@ -42,39 +43,48 @@ class WerkzeugThread(threading.Thread):
 @pytest.fixture(scope='session')
 def werkzeug_server():
     """Run a server from a separate thread."""
-    server = WerkzeugThread()
-    server.start()
-    time.sleep(0.2)
-    yield True
-    server.merge()
+    if CONTEXT.in_github_ci:
+        yield True
+    else:
+        server = WerkzeugThread()
+        server.start()
+        time.sleep(0.2)
+        yield True
+        server.merge()
 
 
 @pytest.fixture(scope='session')
 def gunicorn_server():
     """Run a server from a subprocess."""
-    command = ('python -m gunicorn friends.wsgi:app -w 1 --bind {BIND}'.
-               format(BIND=GUNICORN_OPTIONS['bind']))
-    gunicorn_process = subprocess.Popen(
-        command.split(),
-        cwd=MANAGER_WORKDIR.as_posix(),
-        shell=False,
-    )
-    yield True
-    gunicorn_process.terminate()
+    if CONTEXT.in_github_ci:
+        yield True
+    else:
+        command = ('python -m gunicorn friends.wsgi:app -w 1 --bind {BIND}'.
+                   format(BIND=GUNICORN_OPTIONS['bind']))
+        gunicorn_process = subprocess.Popen(
+            command.split(),
+            cwd=MANAGER_WORKDIR.as_posix(),
+            shell=False,
+        )
+        yield True
+        gunicorn_process.terminate()
 
 
 @pytest.fixture(scope='session')
 def nginx_server():
     """Run a server from a subprocess."""
-    subprocess.call(
-        [OSDIR['sudo'], OSDIR['nginx']],
-        cwd=MANAGER_WORKDIR.as_posix(),
-        shell=False,
-    )
-    time.sleep(1)
-    yield True
-    subprocess.call(
-        [OSDIR['sudo'], OSDIR['nginx'], '-s', 'quit'],
-        cwd=MANAGER_WORKDIR.as_posix(),
-        shell=False,
-    )
+    if CONTEXT.in_github_ci:
+        yield True
+    else:
+        subprocess.call(
+            [OSDIR['sudo'], OSDIR['nginx']],
+            cwd=MANAGER_WORKDIR.as_posix(),
+            shell=False,
+        )
+        time.sleep(1)
+        yield True
+        subprocess.call(
+            [OSDIR['sudo'], OSDIR['nginx'], '-s', 'quit'],
+            cwd=MANAGER_WORKDIR.as_posix(),
+            shell=False,
+        )
